muted_users: Introduce a "Hide" button for revealed messages.
We get rid of the message action popover menu
item and instead switch to a "Hide" button
to hide a revealed message sent by a muted user.
