+++ web/templates/popovers/message_actions_popover.hbs
@@ -67,7 +67,7 @@
         </li>
         {{/if}}
         {{!-- Group 4 --}}
-        {{#if (or should_display_mark_as_unread should_display_remind_me_option should_display_hide_option should_display_collapse should_display_uncollapse)}}
+        {{#if (or should_display_mark_as_unread should_display_remind_me_option should_display_collapse should_display_uncollapse)}}
         <li role="separator" class="popover-menu-separator"></li>
         {{/if}}
         {{#if should_display_mark_as_unread}}
@@ -84,14 +84,6 @@
             <a role="menuitem" data-message-id="{{message_id}}" class="message-reminder popover-menu-link" tabindex="0">
                 <i class="popover-menu-icon zulip-icon zulip-icon-alarm-clock" aria-hidden="true"></i>
                 <span class="popover-menu-label">{{t "Remind me about this" }}</span>
-            </a>
-        </li>
-        {{/if}}
-        {{#if should_display_hide_option}}
-        <li role="none" class="link-item popover-menu-list-item">
-            <a role="menuitem" data-message-id="{{message_id}}" class="rehide_muted_user_message popover-menu-link" tabindex="0">
-                <i class="popover-menu-icon zulip-icon zulip-icon-hide" aria-hidden="true"></i>
-                <span class="popover-menu-label">{{t "Hide muted message again" }}</span>
             </a>
         </li>
         {{/if}}
